extract type utils to its own file
