+++ packages/core/src/types/videoRoom.ts
@@ -1,19 +1,6 @@
 import { SwEvent } from '.'
+import { CamelToSnakeCase, EntityUpdated, VideoEventToInternal } from './utils'
 import { InternalVideoMember } from './videoMember'
-
-export type CamelToSnakeCase<S extends string> =
-  S extends `${infer T}${infer U}`
-    ? `${T extends Capitalize<T>
-        ? '_'
-        : ''}${Lowercase<T>}${CamelToSnakeCase<U>}`
-    : S
-
-export type EntityUpdated<T> = T & {
-  // TODO: `updated` should includes only the "updatable" keys
-  updated: Array<keyof T>
-}
-
-export type VideoEventToInternal<T extends string> = `video.${T}`
 
 /**
  * Public event types
